requirements.txt
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
python-telegram-bot==20.7
flask==3.0.0